requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
amqp==5.1.1
anyio==3.7.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.0
defusedxml==0.7.1
dhooks==1.1.4
Django==4.2.1
dnspython==2.3.0
ElasticEmail==4.0.23
et-xmlfile==1.1.0
eventlet==0.33.3
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
imageio==2.30.0
install==1.3.5
Jinja2==3.1.2
joblib==1.2.0
keepa==1.3.6
kombu==5.3.0
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg==3.1.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
SQLAlchemy==2.0.16
sqlparse==0.4.4
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
